refactor: Removed unused code
