+++ siege-bot/src/commands/all_maps.rs
@@ -12,7 +12,7 @@ use strum::IntoEnumIterator;
 use crate::{formatting::FormatEmbedded, SiegeApi};
 
 use super::{
-    command::DiscordAppCmd, context::DiscordContext, AddUserOptionToCommand, CmdResult,
+    context::DiscordContext, discord_app_command::DiscordAppCmd, AddUserOptionToCommand, CmdResult,
     CommandHandler,
 };
