+++ siege-bot/src/commands/all_operators.rs
@@ -11,7 +11,7 @@ use siege_api::{game_models::Side, models::OperatorStatistics};
 use crate::{formatting::FormatEmbedded, SiegeApi};
 
 use super::{
-    command::DiscordAppCmd, context::DiscordContext, AddUserOptionToCommand, CmdResult,
+    context::DiscordContext, discord_app_command::DiscordAppCmd, AddUserOptionToCommand, CmdResult,
     CommandHandler,
 };
