+++ siege-bot/src/commands/operator.rs
@@ -12,7 +12,7 @@ use siege_api::operator::Operator;
 use crate::{commands::CommandError, constants::NAME, formatting::FormatEmbedded, SiegeApi};
 
 use super::{
-    command::DiscordAppCmd, context::DiscordContext, AddUserOptionToCommand, CmdResult,
+    context::DiscordContext, discord_app_command::DiscordAppCmd, AddUserOptionToCommand, CmdResult,
     CommandHandler,
 };
